add trustline step after deposit has been made
